pyproject.toml
[tool.poetry]
name = "cbproject1"
version = "0.1.0"
description = "cb"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
accelerate = "0.23.0"
aiohttp = "3.8.5"
aiosignal = "1.3.1"
annotated-types = "0.5.0"
antiorm = "1.2.1"
anyio = "3.7.1"
asttokens = "2.4.0"
async-timeout = "4.0.3"
attrs = "23.1.0"
backcall = "0.2.0"
backoff = "2.2.1"
bcrypt = "4.0.1"
beautifulsoup4 = "4.12.2"
bert-score = "0.3.13"
bitsandbytes = "0.41.1"
cachetools = "5.3.1"
ccrawl = "1.9"
certifi = "2022.12.7"
chardet = "5.2.0"
charset-normalizer = "2.1.1"
chroma-hnswlib = "0.7.3"
chromadb = "0.4.14"
click = "8.1.7"
cmake = "3.25.0"
coloredlogs = "15.0.1"
comm = "0.1.4"
contourpy = "1.1.1"
cycler = "0.11.0"
dataclasses-json = "0.5.14"
datasets = "2.14.5"
db = "0.1.1"
db-sqlite3 = "0.0.1"
debugpy = "1.8.0"
decorator = "5.1.1"
dill = "0.3.7"
diskcache = "5.6.3"
dnspython = "2.4.2"
emoji = "2.8.0"
evaluate = "0.4.1"
exceptiongroup = "1.1.3"
executing = "1.2.0"
fastapi = "0.103.2"
filelock = "3.9.0"
filetype = "1.2.0"
flatbuffers = "23.5.26"
fonttools = "4.42.1"
frozenlist = "1.4.0"
fsspec = "2023.6.0"
gguf = "0.3.2"
grandalf = "0.8"
greenlet = "2.0.2"
grpcio = "1.59.0"
h11 = "0.14.0"
httptools = "0.6.0"
huggingface-hub = "0.17.1"
humanfriendly = "10.0"
idna = "3.4"
importlib-metadata = "6.8.0"
importlib-resources = "6.0.1"
ipykernel = "6.25.2"
ipython = "8.15.0"
jedi = "0.19.0"
Jinja2 = "3.1.2"
joblib = "1.3.2"
jsonpatch = "1.33"
jsonpointer = "2.4"
jupyter-client = "8.3.1"
jupyter-core = "5.3.1"
kiwisolver = "1.4.5"
langchain = "0.0.312"
langdetect = "1.0.9"
langsmith = "0.0.43"
libclang = "14.0.1"
lit = "15.0.7"
llama-cpp-python = "0.2.11"
loralib = "0.1.2"
lxml = "4.9.3"
MarkupSafe = "2.1.2"
marshmallow = "3.20.1"
matplotlib = "3.8.0"
matplotlib-inline = "0.1.6"
monotonic = "1.6"
moverscore = "1.0.3"
mpmath = "1.2.1"
multidict = "6.0.4"
multiprocess = "0.70.15"
mypy-extensions = "1.0.0"
nest-asyncio = "1.5.7"
networkx = "3.0"
nltk = "3.8.1"
numexpr = "2.8.6"
# numpy==1.26.0 (commented out because you might not need it)
nvidia-ml-py = "12.535.108"
nvitop = "1.3.1"
onnxruntime = "1.20.1"
overrides = "7.4.0"
packaging = "23.1"
pandas = "2.1.0"
parso = "0.8.3"
peft = "0.5.0"
pexpect = "4.8.0"
pickleshare = "0.7.5"
Pillow = "9.3.0"
platformdirs = "3.10.0"
portalocker = "2.8.2"
posthog = "3.0.2"
prompt-toolkit = "3.0.39"
protobuf = "4.24.4"
psutil = "5.9.5"
ptyprocess = "0.7.0"
pulsar-client = "3.3.0"
pure-eval = "0.2.2"
pyarrow = "13.0.0"
pydantic = "2.3.0"
pydantic-core = "2.6.3"
pyemd = "1.0.0"
Pygments = "2.16.1"
pymongo = "4.5.0"
pyparsing = "3.1.1"
pypdf = "3.16.4"
PyPika = "0.48.9"
pysqlite3-binary = "0.5.2.post1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-iso639 = "2023.6.15"
python-magic = "0.4.27"
python-multipart = "0.0.6"
python-rapidjson = "1.11"
pytz = "2023.3.post1"
PyYAML = "6.0.1"
pyzmq = "25.1.1"
rapidfuzz = "3.4.0"
regex = "2023.8.8"
requests = "2.28.1"
responses = "0.18.0"
safetensors = "0.3.3"
scikit-learn = "1.3.1"
scipy = "1.11.3"
sentence-transformers = "2.2.2"
sentencepiece = "0.1.99"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.5"
SQLAlchemy = "2.0.20"
stack-data = "0.6.2"
starlette = "0.27.0"
sympy = "1.11.1"
tabulate = "0.9.0"
tenacity = "8.2.3"
termcolor = "2.3.0"
threadpoolctl = "3.2.0"
tinydb = "4.8.0"
tokenizers = "0.13.3"
tornado = "6.3.3"
tqdm = "4.66.1"
traitlets = "5.10.0"
transformers = "4.33.2"
triton = "2.0.0"
typer = "0.9.0"
typing = "3.7.4.3"
typing-extensions = "4.8.0"
typing-inspect = "0.9.0"
tzdata = "2023.3"
unstructured = "0.10.21"
urllib3 = "1.26.13"
uvicorn = "0.23.2"
uvloop = "0.17.0"
warcio = "1.7.4"
watchfiles = "0.20.0"
wcwidth = "0.2.6"
websockets = "11.0.3"
xxhash = "3.4.1"
yarl = "1.9.2"
zipp = "3.16.2"
numpy = "1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
